+++ galacteek/config/cmods/app.py
@@ -1,0 +1,6 @@
+from galacteek.config import cGet
+
+
+def defaultContentLangTag():
+    return cGet('defaultContentLanguage',
+                mod='galacteek.application')