+++ galacteek/ld/__init__.py
@@ -13,7 +13,7 @@ def ipsContextUri(contextName: str, ips='galacteek.ld'):
     return URIRef(f'ips://{ips}/{contextName}')
 
 
-def ipsTermUri(name: str, ips='galacteek.ld', fragment=None):
+def ipsTermUri(name: str, fragment: str = None, ips='galacteek.ld'):
     if fragment:
         return URIRef(f'ips://{ips}/{name}#{fragment}')
     else: