+++ galacteek/services/ld/pronto/__models__.py
@@ -1,5 +1,6 @@
 from rdflib import Literal
 
+from galacteek.config.cmods import app as config_app
 from galacteek.core.models.sparql.hashmarks import LDHashmarksSparQLListModel
 from galacteek.core.models.sparql.hashmarks import LDHashmarksItemModel
 from galacteek.core.models.sparql.tags import TagsSparQLModel
@@ -8,10 +9,15 @@ from galacteek.core.models.sparql.tags import TagsPreferencesModel
 
 class ProntoServiceModels:
     async def initializeSparQLModels(self):
+        langTag = config_app.defaultContentLangTag()
+
         # All tags
         self.allTagsModel = TagsSparQLModel(
             graphUri='urn:ipg:i:love:itags',
-            bindings={'tagNameRegex': Literal('')}
+            bindings={
+                'langTag': Literal(langTag),
+                'tagNameRegex': Literal('')
+            }
         )
         self.allTagsModel.update()